requirements.txt
python-dotenv
fastapi
uvicorn[standard]
langgraph>=0.2
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
pydantic>=2
orjson
requests
gdown
numpy
torch
RestrictedPython
reportlab
python-multipart
streamlit
httpx
pytest
